@@ -35,12 +35,12 @@
     "nps": "^5.10.0",
     "null-loader": "^4.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-helmet": "^6.0.0-beta.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-on-rails": "file:.yalc/react-on-rails",
     "react-proptypes": "^1.0.0",
-    "react-redux": "^7.2.2",
+    "react-redux": "^8.0.2",
     "react-router-dom": "^5.2.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.2.0",
@@ -49,7 +49,7 @@
     "sass": "^1.43.4",
     "sass-loader": "^12.3.0",
     "sass-resources-loader": "^2.1.0",
-    "shakapacker": "^6.4.0",
+    "shakapacker": "^6.4.1",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "5",
     "url-loader": "^4.0.0",
@@ -60,6 +60,9 @@
   },
   "devDependencies": {
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/react-helmet": "^6.1.5",
     "react-refresh": "^0.11.0",
     "webpack-dev-server": "^4.7.3"
   },
